+++ mobile/android/android-components/components/concept/engine/src/main/java/mozilla/components/concept/engine/webextension/WebExtensionDelegate.kt
@@ -49,13 +49,6 @@ interface WebExtensionDelegate {
     fun onReady(extension: WebExtension) = Unit
 
     /**
-     * Invoked when optional permissions for a web extension have changed.
-     *
-     * @param extension The [WebExtension] for which permissions have changed.
-     */
-    fun onOptionalPermissionsChanged(extension: WebExtension) = Unit
-
-    /**
      * Invoked when a web extension in private browsing allowed is set.
      *
      * @param extension the modified [WebExtension] instance.