+++ mobile/android/geckoview/src/androidTest/java/org/mozilla/geckoview/test/TranslationsTest.kt
@@ -6,6 +6,7 @@ package org.mozilla.geckoview.test
 
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import androidx.test.filters.MediumTest
+import junit.framework.TestCase.assertFalse
 import junit.framework.TestCase.assertTrue
 import org.json.JSONObject
 import org.junit.After
@@ -78,7 +79,7 @@ class TranslationsTest : BaseSessionTest() {
                 handled.complete(null)
             }
         })
-        var expectedTranslateEvent = JSONObject(
+        val expectedTranslateEvent = JSONObject(
             """
             {
             "actor":{
@@ -89,6 +90,7 @@ class TranslationsTest : BaseSessionTest() {
                       "docLangTag": "es"
                     },
                     "requestedTranslationPair": null,
+                    "hasVisibleChange": false,
                     "error": null,
                     "isEngineReady": false
                     }
@@ -640,5 +642,74 @@ class TranslationsTest : BaseSessionTest() {
                 )
             }
         }
+    }
+
+    @Test
+    fun hasVisibleChangeTest() {
+        mainSession.loadTestPath(TRANSLATIONS_ES)
+        mainSession.waitForPageStop()
+
+        val handled = GeckoResult<Void>()
+        var delegateCalled = 0
+        sessionRule.delegateUntilTestEnd(object : Delegate {
+            @AssertCalled(count = 2)
+            override fun onTranslationStateChange(
+                session: GeckoSession,
+                translationState: TranslationState?,
+            ) {
+                delegateCalled++
+
+                if (delegateCalled == 1) {
+                    assertFalse("Initially not visibly changed.", translationState!!.hasVisibleChange)
+                }
+
+                if (delegateCalled == 2) {
+                    assertTrue("After a translation, the DOM should be visibly changed.", translationState!!.hasVisibleChange)
+                    handled.complete(null)
+                }
+            }
+        })
+        val notTranslated = JSONObject(
+            """
+            {
+            "actor":{
+                "languageState":{
+                    "detectedLanguages": {
+                      "userLangTag": "en",
+                      "isDocLangTagSupported": true,
+                      "docLangTag": "es"
+                    },
+                    "requestedTranslationPair": null,
+                    "hasVisibleChange": false,
+                    "error": null,
+                    "isEngineReady": false
+                    }
+                }
+            }
+            """.trimIndent(),
+        )
+        mainSession.triggerLanguageStateChange(notTranslated)
+
+        val translated = JSONObject(
+            """
+            {
+            "actor":{
+                "languageState":{
+                    "detectedLanguages": {
+                      "userLangTag": "en",
+                      "isDocLangTagSupported": true,
+                      "docLangTag": "es"
+                    },
+                    "requestedTranslationPair": {"fromLanguage" : "es" , "toLanguage" : "en"},
+                    "hasVisibleChange": true,
+                    "error": null,
+                    "isEngineReady": true
+                    }
+                }
+            }
+            """.trimIndent(),
+        )
+        mainSession.triggerLanguageStateChange(translated)
+        sessionRule.waitForResult(handled)
     }
 }